--- packages/oi4-oec-service-node/src/__fixtures__/README
@@ -1,3 +1,0 @@
-DIE Zertifikate SIND 1 JAHR GÜLTIG.
-Die Zertifikate werden also nach einem Jahr ablaufen und können danach für reale requests nicht verwendet werden.
-
