--- packages/oi4-oec-service-node/src/__fixtures__/ca-root-cert.crt
@@ -1,20 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDMzCCAhsCFGfzVfhacEmZSPFSAyrnc22rCwGQMA0GCSqGSIb3DQEBCwUAMFYx
-CzAJBgNVBAYTAmRlMQswCQYDVQQIDAJidzELMAkGA1UEBwwCZnIxCzAJBgNVBAoM
-AmVoMQwwCgYDVQQLDANpb3QxEjAQBgNVBAMMCWZha2UgYXV0aDAeFw0yMjA1MzAx
-MjAyMzFaFw0yMzA1MzAxMjAyMzFaMFYxCzAJBgNVBAYTAmRlMQswCQYDVQQIDAJi
-dzELMAkGA1UEBwwCZnIxCzAJBgNVBAoMAmVoMQwwCgYDVQQLDANpb3QxEjAQBgNV
-BAMMCWZha2UgYXV0aDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAJog
-6YppP2iOm5sj8p017YOxyVFMpBDk9J6K3ZQTA9GAUv02v2HGPqMP/BJIw//jxuEW
-o/SfVrs/1nWqOQNndHvqrG9ZHrgnzzvAiP0QPkuXXjO8QNTeg2iUMxcUUTBuR/FR
-yD6E467d3cS0Rxtw+PYDoNR32VlnJSW/n/jRK4EWI4e6FD2anDkYJD4DFz6WPwf6
-hmnG1u2ULxFncypannWwivy0IUcfKuJBwql7Krk7L4kjqiTpkjzScqAsBr8LSQCJ
-covjOL+QuAGmezOVsObhUy3qfNbL7lqVS+CDfaTgh/L9VBUVAKCqNSaf9/Q8HCEa
-7bCBsMuV3ufLSoQY1H8CAwEAATANBgkqhkiG9w0BAQsFAAOCAQEAkFAciBSLH+Sg
-DQV6jrweCBrUtpcIGj04DJJoEbp3FEiVlk/AG933UVfu4JkYey0Ze5zpsyMEz8V4
-Xl0mbLa4OjReVKtsi3oKpGVw++y/t65SUuubwcv82zmGtOOdFo17ldqnMYmFYLz3
-fxpDwhfJ9VnGzXhBkOdjafXisglYU8Eu+CAZZRVcR5sCEghdZwKHObX2l0JjLg/a
-t+aAMOqp6owBVxjrsK/C+ep5qlGiqBQn0CQ6Ykl5iwbd8TuRkQUWzIOot1c1ivkq
-P9BqYkqrZxMdbyQNU5VuG2vEh+Y/EoBzh+tdR/juXMz0YFTeZ3Rctbdt130Im95Z
-QIE8qOYREg==
------END CERTIFICATE-----
