--- packages/oi4-oec-service-node/src/__fixtures__/decrypted-client-cert.crt
@@ -1,20 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDMzCCAhsCFAE6BxVIYpY7LgRWdlw49C/MJr/RMA0GCSqGSIb3DQEBCwUAMFYx
-CzAJBgNVBAYTAmRlMQswCQYDVQQIDAJidzELMAkGA1UEBwwCZnIxCzAJBgNVBAoM
-AmVoMQwwCgYDVQQLDANpb3QxEjAQBgNVBAMMCWZha2UgYXV0aDAeFw0yMjA1MzAx
-MzM5MTBaFw0yMzA1MjUxMzM5MTBaMFYxCzAJBgNVBAYTAmRlMQswCQYDVQQIDAJi
-dzELMAkGA1UEBwwCZnIxCzAJBgNVBAoMAmVoMQwwCgYDVQQLDANpb3QxEjAQBgNV
-BAMMCWxvY2FsaG9zdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBALXw
-4ChrhccgmsB4ZySTiGBNqiAEIUuyV0Vp3fqcGF8knczuSDc3YjF3YKnDIRRXFwLQ
-ISSvSjSU6WMBQSGH5GaEwjjLRrQfnXD8fguNNLtfI8wC0U1FiF/xixTPmhWj0eiK
-82B+sz3SDnb1WvtOtlwgCGUdLuSzwVW6jORnPGrP2r3QsAMyxbs2RYEKyR4aQfEI
-BK6RBLhXWUhRTuC1/jAmv1rQ9gHkF7HJ0n4uDbet24lUyccAvlx3iMtuB1g0hJ4H
-KWq0cfL702xoIxVVYBoPTc12HEVjve41I45uOQotm7PvMubdQ/+xo99Xprv7OpkF
-tWeAvB2ER7ofaDvLG+UCAwEAATANBgkqhkiG9w0BAQsFAAOCAQEAduzscOS/06cx
-PU32b9ucDax6V8yGTatydbAZKoho3frtcQQZ9CgKSFVgGQeLeiUqNwn+z4TkJEkM
-R6RgJhvnCbs38m5tULPNqFFC8j/9TNXUcOFyuKXq8uLxnlWXA3U8t28mdQ2sEo8c
-4Q9hzFsUbahCh6fFiiFijzspR5VO9IwCn5xnVNo/3O+Df9JMkM5xMccSoI/97cya
-anNwiX5eTzkGN3g3CCjw0Vx4B7GGHjVTe9QUbd90RNzeb0EwrBvzME5Xrvi4YdF3
-h2ZOQYGCOs5GGasx8jkIIcedj6wdwiQVl/HnO9wjuCjrQGcln/5DSJ72m3Otk+1n
-BfdMpPQNbw==
------END CERTIFICATE-----
