--- packages/oi4-oec-service-node/src/__fixtures__/decrypted-client-key.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQC18OAoa4XHIJrA
-eGckk4hgTaogBCFLsldFad36nBhfJJ3M7kg3N2Ixd2CpwyEUVxcC0CEkr0o0lOlj
-AUEhh+RmhMI4y0a0H51w/H4LjTS7XyPMAtFNRYhf8YsUz5oVo9HoivNgfrM90g52
-9Vr7TrZcIAhlHS7ks8FVuozkZzxqz9q90LADMsW7NkWBCskeGkHxCASukQS4V1lI
-UU7gtf4wJr9a0PYB5BexydJ+Lg23rduJVMnHAL5cd4jLbgdYNISeBylqtHHy+9Ns
-aCMVVWAaD03NdhxFY73uNSOObjkKLZuz7zLm3UP/saPfV6a7+zqZBbVngLwdhEe6
-H2g7yxvlAgMBAAECggEAA93/1bwbVtUxlN+6SYO9RVbBQ/bygjpjMn2Ck14YAoWz
-uPRjopDRzH1eF52nKViEi7HC0pXLPPzhTRJswPqzdx5hRnXDpfZ6ETm8sCArO8RU
-7XQt4LeWo3tt2l5jRwx5XLj/XVC0HJ9HYW3rpiXRKHoh04xgP8o2VM0nS3+KQDbr
-lvDuptBcjx7txj4PR/D4HF12Yl5sVJZYSf8v6qy61gL65YnHZDPe8TJJBSvZXlZK
-lyJf+JuaNUttNdxdIZjI+5F+FNE8xgGpy3bz4rx0itl1yQNuqARKrVAtjtLCGS9F
-bUqcnWnluVaW1Ny5zwFuvPZl4+hW5r65ZN49wxmaoQKBgQDEK54KDAnrn06tDvcR
-y5SMy+ETc3ccl8BkE/tCCxeMqSCyGR3/GQcM8FAS4qRnCRLsvV+gWQAc5riXj19B
-NNtJ1j+CIS6j3Z/WOa6wrUz2/q8/gbeRmpkyc+Ic3ArEbC1gsV7nHLT63Cl1FjCa
-SE/NQdpuXc/JVQRPyg7JHvM2jQKBgQDtbkP/gtOx9QnMDiIjYenbIBW2oydgrrVN
-uMBJ35uXzORwCGusHb4kGr+6MbTsTIgEZ3BFvAB7Eda9imyvW2pl5x5Gv3LEe3Vi
-/I6QDS9R5SksiCqppI1nWLhPeM76mbTFGy66aw0PweOsCb1BD6ayLxSzOcxdua7O
-S6cErZdwuQKBgCiAhOm4IJg4VQsgCtzutmX3dioI7e/GWR+QjJDPL3aIOMrsqqfP
-qAgA2dd8Wcs4MWMVmlfF6GMPM89+94DEA/6TsJsSdMm6aVTWpf2uWjSo0s/+3UK/
-s5+aE3Y37DUvzJKpLFn1MFyW9ixJf5FrGqbbUNj2n38HAZPjuPfHXxnFAoGAFvgH
-0/wQBrMx02+6lbmvKeecsiJueMft+irZT0w2fSTxJHVsFSZ7cnqz+Ys9eW6d6e8n
-yVc6pK1BHOz7PquVS25kW2vB6ogW/xB3qi/cWAvXQ9uYZPjqRQksK2QAgfXEqxzK
-FHauuMuE/uuiR10AP1SWqA/7WUXRR1vCOI2ijOECgYBtigkqnrmTBTnpNvSyO45a
-OusWhIk4NQlGE8RLQ8S45uDJGKu3Up8ullofWc+8tWpA95KKuEOAabcMhiTdgzJZ
-/jjsA0QoaiGV3qqyT7lFunCCoUTxmX2vbFgxkyGNFvH341/3vUQu5oS5kdvD2XPq
-NjxHJmKCpf+/fqd0TYIykg==
------END PRIVATE KEY-----
