--- packages/oi4-oec-service-node/src/__fixtures__/encrypted-client-cert.crt
@@ -1,20 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDMzCCAhsCFARB5/82+1o7Yk2ABcDHMp1vALLvMA0GCSqGSIb3DQEBCwUAMFYx
-CzAJBgNVBAYTAmRlMQswCQYDVQQIDAJidzELMAkGA1UEBwwCZnIxCzAJBgNVBAoM
-AmVoMQwwCgYDVQQLDANpb3QxEjAQBgNVBAMMCWZha2UgYXV0aDAeFw0yMjA1MzAx
-MjAzMDRaFw0yMzA1MjUxMjAzMDRaMFYxCzAJBgNVBAYTAmRlMQswCQYDVQQIDAJi
-dzELMAkGA1UEBwwCZnIxCzAJBgNVBAoMAmVoMQwwCgYDVQQLDANpb3QxEjAQBgNV
-BAMMCWxvY2FsaG9zdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAK8F
-XLigxSdkL0pCH4MCPia+7vrhkxZ2Uczz4HaS80dO6WSiG0DQcXUiIkNU4oZw2E7+
-o49z63ycTUf+E1Ac/mPPF3Yf4p2gt2xvfI2kNVolus6GiL3Aly2MJ3t/G1FFaX+O
-MJ9ndn5hVn/9p8D45ckik4dFwjkgc4a8S4TpqVQ+LuARExWdNau0VZuIxE3aO+HS
-UXVk15L64m6aag/+Z4qzhIbTUccNPriJXkG6IcApvZ4JA298Soaqny1apJJaj2M9
-7rouEEwGIYFump26rkf7zlDdZgk4fynXOKDAUNn3h4BGu6U7SH6lxZ+K/kYPTo9Z
-EcF6aruS60MMAcCX628CAwEAATANBgkqhkiG9w0BAQsFAAOCAQEAYWIWnKCSNw2W
-CBtxuzJqesyUFlIGHyt89S2I81GLgHyJbfaA1gY2GqRDgp97Yb9q+uf8KYGoVnJw
-t6e7f7scpMzi+B0gC4TOfhb2vVg5VN114e3WEuP+gcy3U881P2tr99OlHygrch+L
-AUOG1PmOW17QIb5HWl+8+Ymdj6glljKt2QryZfnNn6XjQylb6rSyFH3d8WBv5CiE
-ZzPI463pQ5MIjfKrQZg4EPmVFpPjmw4eoDRzFeUAwnghK8+BvMo9CNM6VbdJOSks
-lO2deqtb1mlRtRl8Clsbwm511Du2b+ndpIPzDwM96zCYw2Ehp4/8WV8mwKY1Shr2
-a0jL8+A2dQ==
------END CERTIFICATE-----
