--- packages/oi4-oec-service-node/src/__fixtures__/encrypted-client-key.key
@@ -1,30 +1,0 @@
------BEGIN ENCRYPTED PRIVATE KEY-----
-MIIFLTBXBgkqhkiG9w0BBQ0wSjApBgkqhkiG9w0BBQwwHAQIQ5OcXJmX2tICAggA
-MAwGCCqGSIb3DQIJBQAwHQYJYIZIAWUDBAEqBBCiNqMlnlMw5S5GajGjfLGpBIIE
-0Gz3Xqj+iB27JG1SZ2zuJKhx0J40/zNhWGzhgW4CHza3rYojQ3QseqMsUZAL+0S+
-FV0sgi+HDPmWy3RNPQsqY3C+kM6ZIpOwAgkYHNSpuPk5OtrP3YagE/DEgVWWVPSQ
-2xMMJvWTXog81dl5eCzZHKM9ac/wNYNk/z4oMaAw81AZ6rlZmBeNYEebDU0+n9IP
-SIN6QgHZTvvx1QlxAdHAhJBF0qIkJ593q5KlHCLDm+8FUdOXuBVgGf8Lq8X0aYig
-XsDPXy1PRjWMH+c/XXwoDlpGqSq/Ef581b7nITgvruhZjmGzVuZ7tZF+EiPlyatU
-umgdkpwRzQnxKVa6oBUjNAmmp0Yev+8yLO09ROO+wXbw3kjQNWw9UtntxT1q9z5n
-Llu1tp+EdRIC2O7C1kssqi8QoszHhwV/rIavVDrKSlHMgDKR9yElEhsAUhmhi5ES
-sQda9yIuKmNojnHsaAxa7fjHgmePT/Uk/qCrrcqR/eISlmHhFmWutiJ5eRPJpImX
-INo/3MWCvQDIlJ8WvsoRzGuGFgSHMzW6ysEflA4Rc3PX58gCGOpAUYOkqGKkeitd
-2McYmwo9SVAeqVqFjLIbArRYlCvXtAAQDjPAxC/4QrwkkjBHrV/2wt7Sf7329FWX
-966XPYccEpMPaajbUOs9og5sZrKBRG23f9spjJVyHmb+3cv7v030KjRdiQT1xYBI
-7j1Yc/JzdolRAl9Zzk/fFKZQDi63i5pQ0nGRq7jnwsXXvahjr+O76ZApJ4WKD+7d
-c4F2fmXvfA7H8la5L5SQFALeFJXqO4C9RB85uGwBoY2PIegIPK5ok2dp/VbNUJHj
-mbuzQyBp5xZE8wT38JFTGd+84NjsLuZILEF3wbQfvcs7cXtYSeLBsX6w35LrxLec
-YNfADi7yzT2ljOz2xTUnaNNyhilW/PU8Sx8LUcT9A0VsrVg6xgLO1h0djD1AHF+b
-2Xb/dP7+CAy/81n5Zqxa1Cdy1qr+eCxO/wwLJEDa8CEvTakJLD85JVNeuSNup7PU
-B/LfxrX8HpdU8PrGR6z40v5226K/QGi24J+ZqKOoyYMbcvZCiJaf1oQNhA+Fbb/2
-rrHkUBuY7XZUxanRZw+S6FcdWkS0UwE+WJUpPfcejFJPjewrZ9ku80/Jm215Hy0T
-Rs9Tt1fcVCalUgp8eWVbhV9FtHu26+9yR3RuX491hYuFYPdpXrC2nhAUWBMhOmkv
-fXkvNakiCadMTDpEcXJRwKnGqxNUIIqgaTtB7C59vNpzIpqRuVg811CQsicN7aAB
-2nZpoQUm3qg+qwJ2QbcZ7BMWkDAFVKYl/UIjlrpc9I1JlxpoxQF/Kx8mVGSDHwqN
-/4hfTih9E2I1GBDO3rqOLmLRYpiqkiv/x8yjMAM4C/oxwMv7DInFheGAOk4yBDzk
-L55o/A1qMu31RRTffoTM7hM3a4Au2Sm8sKbqA25E9paJjjc6ti0in2FxlqJ4t/JA
-mImLtHqdShb2HKi1Kw9upSpIoBs5d8Nc3e6ahBwZzkd4nPXKedDb7q8fJCKPOLWC
-VRhkqDjXvHSnF634TjAQ9CchP/6TOyYi7eXZDORQjwOpgO5SFEhUeKQpf+QvmlUp
-hoSxRdlZe0w5W2KMSUZvZ0MoWd3OWivjJ2i1hDLlfmr3
------END ENCRYPTED PRIVATE KEY-----
