--- packages/oi4-oec-service-node/src/__fixtures__/passphrase.txt
@@ -1,1 +1,0 @@
-abcd
