+++ packages/oi4-oec-service-node/src/index.ts
@@ -1,5 +1,4 @@
-export {OI4MessageBusProxy} from './Proxy/Messagebus';
-export {MqttSettings} from './Proxy/Messagebus/MqttSettings';
+export {OI4MessageBusProxy} from './messageBus';
+export {MqttSettings} from './messageBus/MqttSettings';
 export {ContainerState} from './Container';
-export {OI4WebProxy} from './Proxy/Web';
 export {FileLogger} from './Utilities/FileLogger';
