+++ packages/oi4-oec-service-node/tsconfig.json
@@ -18,8 +18,7 @@
     "src/**/*",
     "src/**/*.json",
     "LICENSE",
-    "README.md",
-    "package.json"
+    "README.md"
   ],
   "exclude": [
     "dist",
